# docs/T033_openapi.yaml
openapi: "3.0.3"
info:
  title: GarudaX Platform API
  description: |
    REST API for the AI Powered Commodity Exchange of Mongolia.
    This gateway translates REST/JSON requests to gRPC calls on backend exchange services.
  version: "1.0.0"
  contact:
    name: GarudaX Platform Engineering
  license:
    name: Proprietary
servers:
  - url: https://api.garudax.mn/api/v1
    description: Production
  - url: https://staging-api.garudax.mn/api/v1
    description: Staging
...
security:
  - BearerAuth: []
paths:
...
  /orders:
    post:
      tags: [Orders]
      summary: Submit a new order
      operationId: submitOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitOrderRequest"
      responses:
        "201":
          description: Order accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionReport"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
    get:
      tags: [Orders]
      summary: List open orders
      operationId: getOpenOrders
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: instrument_id
          in: query
          schema:
            type: string
            format: uuid
        - name: side
          in: query
          schema:
            $ref: "#/components/schemas/Side"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Open orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenOrdersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      tags: [Orders]
      summary: Cancel all orders for an account
      operationId: cancelAllOrders
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: instrument_id
          in: query
          schema:
            type: string
            format: uuid
        - name: side
          in: query
          schema:
            $ref: "#/components/schemas/Side"
      responses:
        "200":
          description: Cancellation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelAllOrdersResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /orders/{order_id}:
    get:
      tags: [Orders]
      summary: Get order details
      operationId: getOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Orders]
      summary: Cancel an order
      operationId: cancelOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Cancellation execution report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionReport"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Orders]
      summary: Modify an order (cancel-replace)
      operationId: modifyOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModifyOrderRequest"
      responses:
        "200":
          description: Modification execution report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionReport"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Market Data ──────────────────────────────────────────────────────
  /instruments/{instrument_id}/book:
    get:
      tags: [Market Data]
      summary: Get L2 order book snapshot
      operationId: getOrderBook
      security: []
      parameters:
        - name: instrument_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: depth
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 50
      responses:
        "200":
          description: Order book snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderBookSnapshot"
...
  /instruments/{instrument_id}/trades/latest:
    get:
      tags: [Market Data]
      summary: Get the most recent trade
      operationId: getLastTrade
      security: []
      parameters:
        - name: instrument_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Most recent trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"

  # ─── Admin ────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
        description: Opaque pagination cursor
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
        X-RateLimit-Limit:
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
        X-RateLimit-Reset:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  reason:
                    type: string
            request_id:
              type: string
          required: [code, message]
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
        has_more:
          type: boolean
    Side:
      type: string
      enum: [buy, sell]

    # ─── Orders ───────────────────────────────────────────────────────
    SubmitOrderRequest:
      type: object
      properties:
        client_order_id:
          type: string
        instrument_id:
          type: string
          format: uuid
        side:
          $ref: "#/components/schemas/Side"
        order_type:
          type: string
          enum: [limit, market, stop_limit, stop_market]
        time_in_force:
          type: string
          enum: [day, gtc, gtd, ioc, fok]
          default: day
        price:
          type: string
          description: Decimal string, required for limit orders
        stop_price:
          type: string
          description: Decimal string, required for stop orders
        quantity:
          type: integer
          minimum: 1
          description: Quantity in lots
        expire_at:
          type: string
          format: date-time
          description: Required for GTD orders
        stp_mode:
          type: string
          enum: [cancel_newest, cancel_oldest, cancel_both]
          default: cancel_newest
      required: [instrument_id, side, order_type, quantity]
    ExecutionReport:
      type: object
      properties:
        exec_id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        client_order_id:
          type: string
        exec_type:
          type: string
          enum: [new, partial_fill, fill, cancelled, rejected, expired, trade_bust]
        order_status:
          type: string
          enum: [new, partially_filled, filled, cancelled, rejected, expired, pending_new]
        side:
          $ref: "#/components/schemas/Side"
        instrument_id:
          type: string
          format: uuid
        price:
          type: string
        quantity:
          type: integer
        last_qty:
          type: integer
        last_price:
          type: string
        cumulative_qty:
          type: integer
        leaves_qty:
          type: integer
        trade_id:
          type: string
          format: uuid
        transact_time:
          type: string
          format: date-time
        reject_reason:
          type: string
        account_id:
          type: string
          format: uuid
    OrderDetail:
      type: object
      properties:
        order_id:
          type: string
          format: uuid
        client_order_id:
          type: string
        instrument_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        participant_id:
          type: string
          format: uuid
        side:
          $ref: "#/components/schemas/Side"
        order_type:
          type: string
          enum: [limit, market, stop_limit, stop_market]
        time_in_force:
          type: string
          enum: [day, gtc, gtd, ioc, fok]
        price:
          type: string
        stop_price:
          type: string
        quantity:
          type: integer
        filled_qty:
          type: integer
        remaining_qty:
          type: integer
        status:
          type: string
          enum: [new, partially_filled, filled, cancelled, rejected, expired, pending_new]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        expire_at:
          type: string
          format: date-time
    ModifyOrderRequest:
      type: object
      properties:
        new_price:
          type: string
          description: New price (decimal string), omit to keep current
        new_quantity:
          type: integer
          description: New quantity, omit or 0 to keep current
    OpenOrdersResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OrderDetail"
        pagination:
          $ref: "#/components/schemas/Pagination"
    CancelAllOrdersResponse:
      type: object
      properties:
        cancelled_count:
          type: integer
        cancelled_order_ids:
          type: array
          items:
            type: string
            format: uuid

    # ─── Market Data ──────────────────────────────────────────────────
    PriceLevel:
      type: object
      properties:
        price:
          type: string
        quantity:
          type: integer
        order_count:
          type: integer
    OrderBookSnapshot:
      type: object
      properties:
        instrument_id:
          type: string
          format: uuid
        bids:
          type: array
          items:
            $ref: "#/components/schemas/PriceLevel"
        asks:
          type: array
          items:
            $ref: "#/components/schemas/PriceLevel"
        last_trade_price:
          type: string
        sequence_number:
          type: integer
        state:
          type: string
          enum: [preopen, auction, continuous, halted, closed]
        timestamp:
          type: string
          format: date-time
...
    Trade:
      type: object
      properties:
        trade_id:
          type: string
          format: uuid
        instrument_id:
          type: string
          format: uuid
        buy_order_id:
          type: string
          format: uuid
        sell_order_id:
          type: string
          format: uuid
        buyer_participant_id:
          type: string
          format: uuid
        seller_participant_id:
          type: string
          format: uuid
        price:
          type: string
        quantity:
          type: integer
        trade_value:
          type: string
        aggressor_side:
          $ref: "#/components/schemas/Side"
        trade_type:
          type: string
          enum: [continuous, auction, block, bust, correction]
        sequence_number:
          type: integer
        executed_at:
          type: string
          format: date-time

    # ─── Admin ────────────────────────────────────────────────────────
...